Rewrite wget.py output
