+++ bot/lib/aioget.py
@@ -8,7 +8,4 @@ async def aioget(url, params={}):
         except aiohttp.client_exceptions.InvalidURL:
             response = await session.get(f"https://{url}", params=params)
 
-        if not response.status == 200:
-            return 404
-
         return response
