+++ bot/lib/convert_bytes.py
@@ -1,0 +1,5 @@
+def convert_bytes(num):
+    for x in ['bytes', 'KB', 'MB', 'GB', 'TB']:
+        if num < 1024.0:
+            return f"{round(num, 2)} {x}"
+        num /= 1024.0
